[test runner] Split operations parsing and execution
